libs/nebius/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "langchain-nebius"
version = "0.1.2"
description = "LangChain integration for Nebius AI Studio"
authors = ["Akim Tsvigun <aktsvigun@nebius.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/nebius/langchain-nebius"
keywords = ["langchain", "nebius", "llm", "ai", "embeddings", "retrieval", "rag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
langchain-core = "0.3.35"
openai = "^1.0.0"
langchain_openai = "0.3.5"

[tool.poetry.dev-dependencies]
pytest = "^8.3.4"
black = ">=24.3.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
langchain-tests = "0.3.1"